gqtest {lmtest}	R Documentation

Goldfeld-Quandt Test

Description

Goldfeld-Quandt test against heteroskedasticity.

Usage

gqtest(formula, point = 0.5, fraction = 0,
  alternative = c("greater", "two.sided", "less"),
  order.by = NULL, data = list())

Arguments

formula
a symbolic description for the model to be tested (or a fitted "lm" object).

point
numerical. If point is smaller than 1 it is interpreted as percentages of data, i.e. n*point is taken to be the (potential) breakpoint in the variances, if n is the number of observations in the model. If point is greater than 1 it is interpreted to be the index of the breakpoint.

fraction
numerical. The number of central observations to be omitted. If fraction is smaller than 1, it is chosen to be fraction*n if n is the number of observations in the model.

alternative
a character string specifying the alternative hypothesis. The default is to test for increasing variances.

order.by
Either a vector z or a formula with a single explanatory variable like ~ z. The observations in the model are ordered by the size of z. If set to NULL (the default) the observations are assumed to be ordered (e.g., a time series).

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which gqtest is called from.

Details

The Goldfeld-Quandt test compares the variances of two submodels divided by a specified breakpoint and rejects if the variances differ.

Under H_0 the test statistic of the Goldfeld-Quandt test follows an F distribution with the degrees of freedom as given in parameter.

Examples can not only be found on this page, but also on the help pages of the data sets bondyield, currencysubstitution, growthofmoney, moneydemand, unemployment, wages.

Value

A list with class "htest" containing the following components:

statistic
the value of the test statistic.

parameter
degrees of freedom.

method
a character string indicating what type of test was performed.

alternative
a character string describing the alternative hypothesis.

p.value
the p-value of the test.

data.name
a character string giving the name(s) of the data.

References

S.M. Goldfeld & R.E. Quandt (1965), Some Tests for Homoskedasticity. Journal of the American Statistical Association 60, 539–547

W. Krämer & H. Sonnberger (1986), The Linear Regression Model under Test. Heidelberg: Physica

See Also

lm

Examples

## generate a regressor
x <- rep(c(-1,1), 50)
## generate heteroskedastic and homoskedastic disturbances
err1 <- c(rnorm(50, sd=1), rnorm(50, sd=2))
err2 <- rnorm(100)
## generate a linear relationship
y1 <- 1 + x + err1
y2 <- 1 + x + err2
## perform Goldfeld-Quandt test
gqtest(y1 ~ x)
gqtest(y2 ~ x)

[Package lmtest version 0.9-40 Index]