resettest {lmtest}	R Documentation

RESET Test

Description

Ramsey's RESET test for functional form.

Usage

resettest(formula, power = 2:3, type = c("fitted", "regressor",
  "princomp"), data = list(), vcov = NULL, ...)

Arguments

formula
a symbolic description for the model to be tested (or a fitted "lm" object).

power
integers. A vector of positive integers indicating the powers of the variables that should be included. By default, the test is for quadratic or cubic influence of the fitted response.

type
a string indicating whether powers of the fitted response, the regressor variables (factors are left out), or the first principal component of the regressor matrix should be included in the extended model.

data
an optional data frame containing the variables in the model. By default the variables are taken from the environment which resettest is called from.

vcov, ...
optional arguments to be passed to waldtest for carrying out the F test.

Details

The RESET test is a popular diagnostic for correctness of functional form. The basic assumption is that under the alternative the model can be written in the form y=X * beta + Z * gamma. Z is generated by taking powers either of the fitted response, the regressor variables, or the first principal component of X. A standard F-Test is then applied to determine whether these additional variables have significant influence. The test statistic under H_0 follows an F distribution with parameter degrees of freedom.

This function was called reset in previous versions of the package. Please use resettest instead.

Examples can not only be found on this page, but also on the help pages of the data sets bondyield, currencysubstitution, growthofmoney, moneydemand, unemployment, wages.

Value

An object of class "htest" containing:

statistic
the test statistic.

p.value
the corresponding p-value.

parameter
degrees of freedom.

method
a character string with the method used.

data.name
a character string with the data name.

References

J.B. Ramsey (1969), Tests for Specification Errors in Classical Linear Least-Squares Regression Analysis. Journal of the Royal Statistical Society, Series B 31, 350–371

W. Krämer & H. Sonnberger (1986), The Linear Regression Model under Test. Heidelberg: Physica

See Also

lm

Examples

x <- c(1:30)
y1 <- 1 + x + x^2 + rnorm(30)
y2 <- 1 + x + rnorm(30)
resettest(y1 ~ x, power=2, type="regressor")
resettest(y2 ~ x, power=2, type="regressor")

[Package lmtest version 0.9-40 Index]