coeftest {lmtest}	R Documentation

Inference for Estimated Coefficients

Description

coeftest is a generic function for performing z and (quasi-)t Wald tests of estimated coefficients. coefci computes the corresponding Wald confidence intervals.

Usage

coeftest(x, vcov. = NULL, df = NULL, ...)

coefci(x, parm = NULL, level = 0.95, vcov. = NULL, df = NULL, ...)

Arguments

x
an object (for details see below).

vcov.
a specification of the covariance matrix of the estimated coefficients. This can be specified as a matrix or as a function yielding a matrix when applied to x.

df
the degrees of freedom to be used. If this is a finite positive number a t test with df degrees of freedom is performed. In all other cases, a z test (using a normal approximation) is performed. By default it tries to use x$df.residual and performs a z test if this is NULL.

...
further arguments passed to the methods and to vcov. in the default method.

parm
a specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level
the confidence level required.

Details

The generic function coeftest currently has a default method (which works in particular for "lm" and "glm" objects) and a method for objects of class "breakpointsfull" (as computed by breakpoints.formula).

The default method assumes that a coef methods exists, such that coef(x) yields the estimated coefficients.

To specify a covariance matrix vcov. to be used, there are three possibilities: 1. It is pre-computed and supplied in argument vcov.. 2. A function for extracting the covariance matrix from x is supplied, e.g., vcovHC or vcovHAC from package sandwich. 3. vcov. is set to NULL, then it is assumed that a vcov method exists, such that vcov(x) yields a covariance matrix. For illustrations see below.

The degrees of freedom df determine whether a normal approximation is used or a t distribution with df degrees of freedoms is used. The default method uses df.residual(x) and if this is NULL a z test is performed.

The generic function coefci computes the corresponding Wald confidence intervals.

Value

coeftest returns an object of class "coeftest" which is essentially a coefficient matrix with columns containing the estimates, associated standard errors, test statistics and p values.

coefci returns a matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in percent.

See Also

lm, waldtest

Examples

## load data and fit model
data("Mandible", package = "lmtest")
fm <- lm(length ~ age, data = Mandible, subset=(age <= 28))

## the following commands lead to the same tests:
summary(fm)
coeftest(fm)

## a z test (instead of a t test) can be performed by
coeftest(fm, df = Inf)

## corresponding confidence intervales
coefci(fm)
## which in this simple case is equivalent to
confint(fm)

if(require("sandwich")) {
## a different covariance matrix can be also used:
## either supplied as a function
coeftest(fm, df = Inf, vcov = vcovHC)
## or as a function with additional arguments
coeftest(fm, df = Inf, vcov = vcovHC, type = "HC0")
## or as a matrix
coeftest(fm, df = Inf, vcov = vcovHC(fm, type = "HC0"))
}

[Package lmtest version 0.9-35 Index]